// src/app/models/novel.model.ts
export class Novel {
  createAt?: string;
  lastModifiedAt?: string;
  createdBy?: string;
  lastModifiedBy?: string;
  id?: number;
  title?: string;
  author?: string;
  url?: string;
  imageUrl?: string;
  rating?: number;
  totalVotes?: number;
  language?: string;
  artist?: string;
  releaseYear?: number;
  status?: string;
  publisher?: string;
  description?: string;
  createdAt?: string;
  updatedAt?: string;
  imageData?: string[];
}